==> scripts/00_GitHub_Settings.R <==
install.packages("usethis")
install.packages("gitcreds")

gitcreds::gitcreds_set()

library(usethis)
use_git_config(user.name = "Michal Michalski", user.email = "michalski82@gmail.com")

library(usethis)
use_git_config(user.name = "Michal Michalski", user.email = "michalski82@gmail.com")


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

#https://indrajeetpatil.github.io/ggstatsplot/

st_layers("data/data.gpkg")
